move uri package in to identifier
