+++ packages/identifier/README.md
@@ -43,6 +43,19 @@ NSID.isValid('example.com/foo') // => false
 NSID.isValid('foo') // => false
 ```
 
+## AT URI
+
+```typescript
+import { AtUri } from '@atproto/identifier'
+
+const uri = new AtUri('at://bob.com/com.example.post/1234')
+uri.protocol   // => 'at:'
+uri.origin     // => 'at://bob.com'
+uri.hostname   // => 'bob.com'
+uri.collection // => 'com.example.post'
+uri.rkey       // => '1234'
+```
+
 ## License
 
 MIT
