+++ packages/identifier/src/index.ts
@@ -1,3 +1,4 @@
 export * from './handle'
 export * from './did'
 export * from './nsid'
+export * from './aturi'
